@@ -1,37 +1,10 @@
-/*
- *  File:   MusicModule.cs
- *  Author: Angelo Breuer
- *
- *  The MIT License (MIT)
- *
- *  Copyright (c) Angelo Breuer 2022
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-
 namespace Lavalink4NET.Discord_NET.ExampleBot;
 
 using System;
 using System.Threading.Tasks;
 using Discord.Interactions;
-using Lavalink4NET.Player;
-using Lavalink4NET.Rest;
+using Lavalink4NET.Players;
+using Lavalink4NET.Rest.Entities.Tracks;
 
 /// <summary>
 ///     Presents some of the main features of the Lavalink4NET-Library.
@@ -58,17 +31,17 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     [SlashCommand("disconnect", "Disconnects from the current voice channel connected to", runMode: RunMode.Async)]
     public async Task Disconnect()
     {
-        var player = await GetPlayerAsync();
+        var player = await GetPlayerAsync().ConfigureAwait(false);
 
-        if (player == null)
+        if (player is null)
         {
             return;
         }
 
         // when using StopAsync(true) the player also disconnects and clears the track queue.
         // DisconnectAsync only disconnects from the channel.
-        await player.StopAsync(true);
-        await ReplyAsync("Disconnected.");
+        await player.StopAsync(true).ConfigureAwait(false);
+        await ReplyAsync("Disconnected.").ConfigureAwait(false);
     }
 
     /// <summary>
@@ -79,30 +52,32 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     [SlashCommand("play", description: "Plays music", runMode: RunMode.Async)]
     public async Task Play(string query)
     {
-        var player = await GetPlayerAsync();
+        var player = await GetPlayerAsync().ConfigureAwait(false);
 
-        if (player == null)
+        if (player is null)
         {
             return;
         }
 
-        var track = await _audioService.GetTrackAsync(query, SearchMode.YouTube);
+        var track = await _audioService.Tracks
+            .LoadTrackAsync(query, TrackSearchMode.YouTube)
+            .ConfigureAwait(false);
 
-        if (track == null)
+        if (track is null)
         {
-            await ReplyAsync("😖 No results.");
+            await ReplyAsync("😖 No results.").ConfigureAwait(false);
             return;
         }
 
-        var position = await player.PlayAsync(track, enqueue: true);
+        var position = await player.PlayAsync(track, enqueue: true).ConfigureAwait(false);
 
-        if (position == 0)
+        if (position is 0)
         {
-            await ReplyAsync("🔈 Playing: " + track.Source);
+            await ReplyAsync($"🔈 Playing: {track.Uri}").ConfigureAwait(false);
         }
         else
         {
-            await ReplyAsync("🔈 Added to queue: " + track.Source);
+            await ReplyAsync($"🔈 Added to queue: {track.Uri}").ConfigureAwait(false);
         }
     }
 
@@ -113,20 +88,20 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     [SlashCommand("position", description: "Shows the track position", runMode: RunMode.Async)]
     public async Task Position()
     {
-        var player = await GetPlayerAsync();
+        var player = await GetPlayerAsync().ConfigureAwait(false);
 
-        if (player == null)
+        if (player is null)
         {
             return;
         }
 
-        if (player.CurrentTrack == null)
+        if (player.CurrentTrack is null)
         {
-            await ReplyAsync("Nothing playing!");
+            await ReplyAsync("Nothing playing!").ConfigureAwait(false);
             return;
         }
 
-        await ReplyAsync($"Position: {player.Position.Position} / {player.CurrentTrack.Duration}.");
+        await ReplyAsync($"Position: {player.Position.Position} / {player.CurrentTrack.Duration}.").ConfigureAwait(false);
     }
 
     /// <summary>
@@ -138,19 +113,19 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     {
         var player = await GetPlayerAsync(connectToVoiceChannel: false);
 
-        if (player == null)
+        if (player is null)
         {
             return;
         }
 
-        if (player.CurrentTrack == null)
+        if (player.CurrentTrack is null)
         {
-            await ReplyAsync("Nothing playing!");
+            await ReplyAsync("Nothing playing!").ConfigureAwait(false);
             return;
         }
 
-        await player.StopAsync();
-        await ReplyAsync("Stopped playing.");
+        await player.StopAsync().ConfigureAwait(false);
+        await ReplyAsync("Stopped playing.").ConfigureAwait(false);
     }
 
     /// <summary>
@@ -163,19 +138,19 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     {
         if (volume is > 1000 or < 0)
         {
-            await ReplyAsync("Volume out of range: 0% - 1000%!");
+            await ReplyAsync("Volume out of range: 0% - 1000%!").ConfigureAwait(false);
             return;
         }
 
-        var player = await GetPlayerAsync();
+        var player = await GetPlayerAsync().ConfigureAwait(false);
 
-        if (player == null)
+        if (player is null)
         {
             return;
         }
 
-        await player.SetVolumeAsync(volume / 100f);
-        await ReplyAsync($"Volume updated: {volume}%");
+        await player.setv(volume / 100f).ConfigureAwait(false);
+        await ReplyAsync($"Volume updated: {volume}%").ConfigureAwait(false);
     }
 
     /// <summary>
@@ -189,11 +164,11 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
     /// </returns>
     private async ValueTask<VoteLavalinkPlayer> GetPlayerAsync(bool connectToVoiceChannel = true)
     {
-        var player = _audioService.GetPlayer<VoteLavalinkPlayer>(Context.Guild.Id);
+        var player = await _audioService.Players
+            .GetPlayerAsync<VoteLavalinkPlayer>(Context.Guild.Id)
+            .ConfigureAwait(false);
 
-        if (player != null
-            && player.State != PlayerState.NotConnected
-            && player.State != PlayerState.Destroyed)
+        if (player is null && player.State is not PlayerState.Destroyed)
         {
             return player;
         }
@@ -202,16 +177,18 @@ public sealed class MusicModule : InteractionModuleBase<SocketInteractionContext
 
         if (!user.VoiceState.HasValue)
         {
-            await ReplyAsync("You must be in a voice channel!");
+            await ReplyAsync("You must be in a voice channel!").ConfigureAwait(false);
             return null;
         }
 
         if (!connectToVoiceChannel)
         {
-            await ReplyAsync("The bot is not in a voice channel!");
+            await ReplyAsync("The bot is not in a voice channel!").ConfigureAwait(false);
             return null;
         }
 
-        return await _audioService.JoinAsync<VoteLavalinkPlayer>(user.Guild.Id, user.VoiceChannel.Id);
+        return await _audioService.Players
+            .JoinAsync<VoteLavalinkPlayer>(user.Guild.Id, user.VoiceChannel.Id)
+            .ConfigureAwait(false);
     }
 }
