@@ -8,31 +8,39 @@ using Lavalink4NET.Integrations;
 using Lavalink4NET.Protocol.Payloads;
 using Lavalink4NET.Rest;
 using Lavalink4NET.Socket;
+using Lavalink4NET.Tracks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 public sealed partial class AudioService : IAudioService
 {
     private readonly TaskCompletionSource _readyTaskCompletionSource;
-    private readonly LavalinkSocket _socket;
     private readonly ILogger<AudioService> _logger;
+    private readonly IDiscordClientWrapper _clientWrapper;
+    private readonly ILoggerFactory _loggerFactory;
+    private readonly LavalinkNodeOptions _options;
 
-    public AudioService(IDiscordClientWrapper clientWrapper, ILavalinkApiClient apiClient, ILoggerFactory loggerFactory)
+    public AudioService(
+        IDiscordClientWrapper clientWrapper,
+        ILavalinkApiClient apiClient,
+        ITrackManager trackManager,
+        ILoggerFactory loggerFactory,
+        IOptions<LavalinkNodeOptions> options)
     {
-        // TODO
-        var socketOptions = new LavalinkSocketOptions
-        {
-
-        };
+        ArgumentNullException.ThrowIfNull(clientWrapper);
+        ArgumentNullException.ThrowIfNull(apiClient);
+        ArgumentNullException.ThrowIfNull(trackManager);
+        ArgumentNullException.ThrowIfNull(loggerFactory);
+        ArgumentNullException.ThrowIfNull(options);
 
         Players = new PlayerManager(clientWrapper, apiClient);
+        Tracks = trackManager;
         Integrations = new IntegrationCollection();
 
+        _clientWrapper = clientWrapper;
+        _loggerFactory = loggerFactory;
+        _options = options.Value;
         _logger = loggerFactory.CreateLogger<AudioService>();
-
-        _socket = new LavalinkSocket(
-            logger: loggerFactory.CreateLogger<LavalinkSocket>(),
-            options: Options.Create(socketOptions));
 
         _readyTaskCompletionSource = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
 
@@ -46,6 +54,8 @@ public sealed partial class AudioService : IAudioService
     public IPlayerManager Players { get; }
 
     public IIntegrationCollection Integrations { get; }
+
+    public ITrackManager Tracks { get; }
 
     public void Dispose()
     {
@@ -118,9 +128,57 @@ public sealed partial class AudioService : IAudioService
     {
         cancellationToken.ThrowIfCancellationRequested();
 
+        static async ValueTask<ClientInformation?> WaitForClientReadyAsync(IDiscordClientWrapper clientWrapper, CancellationToken cancellationToken = default)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            ArgumentNullException.ThrowIfNull(clientWrapper);
+
+            using var cancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
+            cancellationTokenSource.CancelAfter(TimeSpan.FromSeconds(10));
+
+            try
+            {
+                return await clientWrapper
+                    .WaitForReadyAsync(cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            catch (OperationCanceledException) when (!cancellationToken.IsCancellationRequested)
+            {
+                return null;
+            }
+        }
+
+        _logger.LogDebug("Waiting for client being ready...");
+        _logger.LogInformation("Lavalink node initialized.");
+
+        var clientInformation = await WaitForClientReadyAsync(
+            clientWrapper: _clientWrapper,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+
+        if (clientInformation is null)
+        {
+            var exception = new TimeoutException("Timed out while waiting for discord client being ready.");
+            _logger.LogError(exception, "Timed out while waiting for discord client being ready.");
+            throw exception;
+        }
+
+        _logger.LogDebug("Discord client ({ClientLabel}) is ready.", clientInformation.Value.Label);
+
+        var socketOptions = new LavalinkSocketOptions
+        {
+            HttpClientName = _options.HttpClientName,
+            Uri = _options.Uri,
+            ShardCount = clientInformation.Value.ShardCount,
+            UserId = clientInformation.Value.CurrentUserId,
+        };
+
+        using var socket = new LavalinkSocket(
+            logger: _loggerFactory.CreateLogger<LavalinkSocket>(),
+            options: Options.Create(socketOptions));
+
         while (true)
         {
-            var payload = await _socket
+            var payload = await socket
                 .ReceiveAsync(cancellationToken)
                 .ConfigureAwait(false);
 
